Updated download script to ease traffic usage for gh

+++ world/download-clouds.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
 
-wget --no-cache https://github.com/apollo-ng/cloudmap/blob/master/global.jpg 
-cp global.jpg clouds.jpg
-rm global.jpg
-#convert -resize 5400x2700 clouds_2048.jpg clouds.jpg
-#rm clouds_2048.jpg
+ORIGINSHA=$(wget https://raw.githubusercontent.com/apollo-ng/cloudmap/master/global.sha256 --no-cache -q -O - | awk {'print $1;'})
+LOCALSHA=$(sha256sum clouds.jpg | awk {'print $1;'})
+
+if [ ${ORIGINSHA} != ${LOCALSHA} ]; 
+then
+    wget -O global.jpg --no-cache -q https://raw.githubusercontent.com/apollo-ng/cloudmap/master/global.jpg?${ORIGINSHA} 
+    
+    NEWSHA=$(sha256sum global.jpg | awk {'print $1;'})
+   
+    if [ $NEWSHA == $ORIGINSHA ];
+    then
+        mv global.jpg clouds.jpg
+    else
+        rm global.jpg
+    fi
+fi
